This update contains one critical bug fix, and one added feature:
CRITICAL BUG FIX: Subroutine collect was being called instead of collect_loc
in some of the NMM INITPOSTs.  That caused the CENLAT to be incorrect when more
than two processors was in use.  

The COLLECT routine should NEVER be called unless all processors have an IMxJM
array that they're using to store their own little patch.  If the per-process
array is allocated as (IM,jsta_2l:jend_2l) then calling collect WILL break 
MPI support.

This is the problem that we (EMC HWRF) earlier mistaked as being isolated to synthetic
satellite output.  It was isolated to that output before because we
used more than two processors only when we were running UPP to create the
synthetic satellite output.

ADDED FEATURE: needed by HWRF, there are two new output variables: 601 and 602,
which are MODEL SFC U WIND STR and MODEL SFC V WIND STR.  Those are the same as
SFC U WIND STRESS (and V) but use the model output TAUX and TAUY instead of
calling CALTAU.  The output kpds values are deliberately the same for the
model and calculated versions.


git-svn-id: https://svn-dtc-unifiedpostproc.cgd.ucar.edu/branches/2011_oper@101 3ad70301-0a67-45a7-a7d4-3cbace1a8cc5