Changes are:
a) Two options d3d_on and gocart_on that Moorthi to POST to reduce memory are hardcoded in current post code, these two option are put in post namelist too, so the code does not need to be recompiled when ngac post is running
b) An option is added to POST namelist to decide if pressure levels P0 in POST defined from post script is in pascal or hpa unit. It is because in grib2 pressure is in unit pascal, and the pressure levels in POST control file is in pascal too, This option p0pascal is added to post namelist
c) WAM post needs to output fields in isentropic levels, the isentropic levels are currently missing in SET_LVLSXML.f, fix has been made to SET_LVLSXML.f, SET_OUTFLDS.f and WRFPOST.f.

M       SET_LVLSXML.f
M       SET_OUTFLDS.f
M       WRFPOST.f