Second stage in removing WRF dependency:
Now contains a new directory "wrf_io/" with its own makefile for building the
relevant WRF code. Now you no longer have to pre-build WRF, only run the
configure script to create a configure.wrf file.

To build this version:

 - Copy in the external/, frame/, inc/, main/, Registry/, and tools/ directories from the RAW UNBUILT WRF source code into the new wrf_io/ directory
 - Copy in a configure.wrf file created with that same WRF source code

Should be able to pare down the above list soon, many of them are only
needed for one or two files (or perhaps ultimately none)

This version was tested with the WRF V4.0 Friendly Release #3 (git hash e96ef955), but it should work for a wide range of WRF code