Revamping how we check for JASPER libraries/include files.
Old code incorrectly and unexpectedly defaulted to a (typically) non-existant location for JASPERLIB and JASPERINC.

In keeping with the 'Let the configure script fail'(TM) philosophy I'm trying to instill here,
if the user does not set JASPER or JASPERLIB and JASPERINC variables, unless one of the old default locations exists,
the configure script will fail with an error message. Users can overwrite this message by setting NOGRIB2 environment
variable, but support for GRIB1 will be going away in the not-so-distant future!