Initialize the AQF related variables to be "0." instead of "spval"
The purpose for the initialization is for run without itag "aqfcmaq_on=.true." (default is "aqfcmaq_on=.false.").  The value of PMTF & OZCON shown in the output will be "0" instead of misled value related to the calculation from "spval".  O3MR is the tracer also in FV3GFS, the process is outside the "aqfcmaq_on" itag control.