Code and parm modification for the RRFS-CMAQ chemistry calciulation (#346)
* New modification for the RRFS-CMAQ PMTF (PM2.5)

New modification for the RRFS-CMAQ

* Modification to include necessary variables for the RRFS-CMAQ

Modification to include necessary variables for the RRFS-CMAQ

* Delete fv3lam_cmaq_post_avblflds.xml-org

* Delete fv3lam_cmaq.xml-org

* Delete fv3lam_cmaq.xml-pmtf_only

* Delete fv3lam_cmaq_post_avblflds.xml-new

* Delete makefile-cmaq

* Delete makefile-org

* Delete ALLOCATE_ALL.f-org

* Delete ALLOCATE_ALL.f-pmtf

* Delete DEALLOCATE.f-org

* Delete DEALLOCATE.f-pmtf

* Delete INITPOST_NETCDF.f-org

* Delete INITPOST_NETCDF.f-pmtf

* Delete MDLFLD.f-org

* Delete MDLFLD.f-pmtf

* Delete MDLFLD.f-pmtf-unit

* Delete RQSTFLD.F-org

* Delete RQSTFLD.F-pmtf

* Delete VRBLS3D_mod.f-org

* Delete VRBLS3D_mod.f-pmtf

* unified "post_avblflds.xml" related change

Update files according to the removal of "fv3lam_cmaq_post_avblflds.xml" to a unified "post_avblflds.xml"

* Delete fv3lam_cmaq_post_avblflds.xml

to an unified "post_avblflds.xml"

* remove variables do not in the GRIB2 TABLE

Update the file for variables used in the output and defined in GRIB2 TABLE

* Update codes to consistent with changes in parm

Remove variables that will not be used in UPP output.

* Remove O3MR

Remove O3MR that has been processed in the original code
The removal of O3MR in both codes also eliminate the duplicated "O3MR" in UPP output

* Change the allocation 

Add "aqfcmaq_on" as the itag for the conditional allocation for the input variables.  Only variables in the output are set in the main allocation, the rest are set as local in the "INITPOST_NETCDF" and will only be used when "aqfcmaq_on=.true."

* Initialize the AQF related variables to be "0." instead of "spval"

The purpose for the initialization is for run without itag "aqfcmaq_on=.true." (default is "aqfcmaq_on=.false.").  The value of PMTF & OZCON shown in the output will be "0" instead of misled value related to the calculation from "spval".  O3MR is the tracer also in FV3GFS, the process is outside the "aqfcmaq_on" itag control.

* renew file

Reload the file to eliminate the "block & spacce" (-b) difference from the code diff

* Further limit the allociation of AQF UPP output

Further limit the allocation of AQF UPP output variables only exist when itag "aqfcmaq_on=.true." (default .false.).  With itag "aqfcmaq=.true." , the UPP output will contains the allocated variables (currently, OZCON, PMTF).  Without the itag (default .false.), the UPP output will return to original non-CMAQ version (no OZCON & PMTF)

* Remove O3MR

Remove O3mr that has been coded in the original code as "O3" from related codes

* change makefile 

comment out the RRFS-CMAQ related process currently for it is not in the RRFS prallel run

* New NT file for chemistry only

file contains only chemical species that have been added to the source code "initpost_netcdf".

* Revert makefile to original

exclude the generation of RRFS-CMAQ control file related changes until parallel setup.