Update MISCLN.f
Removed EL_SCHEME flag and the EL_SCHEME 2 as it was intended as a method of validating the output of EL_SCHEME effective layer top and bottom calculations. The HWT/SFE confirmed that the EL_SCHEME 1 computations were viable making EL_SCHEME 2 superfluous. The debugprint flag was also set to false as it currently not needed in the execution of the MISCLN subprogram, however, it may be switch on at any point it becomes necessary to trace a problem.