Correction to the effective layer top and bottom calculation.  (#358)
* Update params_grib2_tbl_new

* Update post_avblflds.xml

* This commit includes changes to the computation of the effective layer top and bottom.

* This commit includes changes to the computation of the effective layer top and bottom.

* Update ALLOCATE_ALL.f

i loop indices corrected.

* Update MISCLN.f

* Update MISCLN.f

Removed EL_SCHEME flag and the EL_SCHEME 2 as it was intended as a method of validating the output of EL_SCHEME effective layer top and bottom calculations. The HWT/SFE confirmed that the EL_SCHEME 1 computations were viable making EL_SCHEME 2 superfluous. The debugprint flag was also set to false as it currently not needed in the execution of the MISCLN subprogram, however, it may be switch on at any point it becomes necessary to trace a problem.

* Update UPP_PHYSICS.f

Removed the definition of CALCAPE1D which was invoked in MISCLN.f as the CAPE/CIN computational routine used in EL_SCHEME 2. It is now a defunct option.

* Update MISCLN.f

* Update MISCLN.f

* Added conditional block that only invokes the effective layer top and bottom correction for RTMA.

* Addressed remaining bug linked to the use of the FV3R first guess.

* minor code changes made to ALLOCATE_ALL.f, MISCNL.f and UPP_PHYSICS.f to address segmentation fault occurring during the rtma regression tests

* Clean up before merging.